rename page to reply tracker
